InitialDataGenerator and Getter added. TrackingItem renamed to ItemTracking. All ID's has "long" type.

+++ src/main/java/ru/digdes/steammarketparser/model/entity/Price.java
@@ -22,7 +22,7 @@ public class Price {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "PriceID")
-    private int priceID;
+    private long priceID;
 
     @OneToOne
     @JoinColumn(name="ItemID")
